requirements.txt
langchain==0.1.9
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.0.8
langsmith==0.1.147
openai==1.59.8
numpy==1.26.4
streamlit==1.44.1 
pandas==2.2.3
transformers
python-dotenv
torch
openai==1.59.8
httpx
statsmodels